refactor: pkg dataapps/api: create auth_provider/auth_rule files
